S-SOCRATES-BE/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
httpx>=0.27.0
numpy>=1.26.0
faster-whisper>=0.9.0
edge-tts>=6.1.9
llama-index>=0.10.0
llama-index-llms-openai-like>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
sentence-transformers>=2.2.2
llama-index-llms-gemini>=0.1.0
google-generativeai>=0.5.0
google-cloud-texttospeech>=2.16.0
